requirements.txt
langchain
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb